requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
matplotlib==3.10.5
seaborn==0.13.2
pandas==2.3.2
numpy==2.3.3
pytest==8.4.2
httpx==0.28.1
openai==1.101.0
python-dotenv==1.1.0
